% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr_plot.R
\name{autoplot.statement}
\alias{autoplot.statement}
\title{autoplot statement}
\usage{
autoplot.statement(x, ...)
}
\arguments{
\item{x}{statement object}

\item{...}{other params}
}
\description{
autoplot statement
}
\keyword{internal}

