% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{calculate}
\alias{calculate}
\title{Calculate formulas}
\usage{
calculate(x, ...)
}
\arguments{
\item{x}{a statement object}

\item{...}{list of formulas}
}
\description{
Calculate formulas
}
\examples{
\dontrun{

balance_sheet \%>\% calculate(

  current_ratio = AssetsCurrent / LiabilitiesCurrent,

  quick_ratio =
    ( CashAndCashEquivalentsAtCarryingValue +
        AvailableForSaleSecuritiesCurrent +
        AccountsReceivableNetCurrent
      ) / LiabilitiesCurrent
)
}
}

