% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{calculation}
\alias{calculation}
\title{Define calculation}
\usage{
calculation(...)
}
\arguments{
\item{...}{formulas}
}
\description{
Define calculation
}
\seealso{
\code{\link{do_calculation}}
}

