% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{check_statement}
\alias{check_statement}
\title{Check statement}
\usage{
check_statement(statement)
}
\arguments{
\item{statement}{statement object}
}
\value{
check object with calculated and original values
}
\description{
Checks statement calculation consistency
}

