% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{diff.statement}
\alias{diff.statement}
\title{Statement lagged differences}
\usage{
\method{diff}{statement}(x, lag = 1L, ...)
}
\arguments{
\item{x}{a statement to be differenced}

\item{lag}{an integer indicating which lag to use}

\item{...}{further arguments passed to or from other methods}
}
\value{
a statement object equal to successive  differences (x and lagged x)
}
\description{
Statement lagged differences
}

