% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{do_calculation}
\alias{do_calculation}
\title{Run the calculation}
\usage{
do_calculation(x, calculation)
}
\arguments{
\item{x}{statement object}

\item{calculation}{a calculation expression}
}
\description{
Run the calculation
}
\seealso{
\code{\link{calculation}}
}

