% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{expose}
\alias{expose}
\title{Calculate higher order element values}
\usage{
expose(x, ...)
}
\arguments{
\item{x}{a statement object}

\item{...}{elements of the statement}
}
\value{
a statement object
}
\description{
Generate statement with agregated values.
}
\details{
expose uses calculation link base hierarchy to find
which concept to include in the aggregate. The aggregation function is
always \emph{sum}. See examples for more details.
}
\examples{
\dontrun{

# total assets
st1$StatementOfFinancialPositionClassified \%>\%
  expose(
    "Assets"
  )

# current and non-current assets (using other)
st1$StatementOfFinancialPositionClassified \%>\%
  expose(
    CurrentAssets = "AssetsCurrent",
    NoncurrentAssets = other("Assets"),
    "LiabilitiesAndStockholdersEquity"
  )

# calculate non-current assets (using \%without\% operator)
st1$StatementOfFinancialPositionClassified \%>\%
  expose(
    NoncurrentAssets = "Assets" \%without\% "AssetsCurrent",
    CurrentAssets = "AssetsCurrent",
    "LiabilitiesAndStockholdersEquity"
  )
}
}
\seealso{
\link{finstr}
}

