% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{fold}
\alias{fold}
\title{Collapses columns to specified groups}
\usage{
fold(x, element_groups)
}
\arguments{
\item{x}{a statement object}

\item{element_groups}{A list of groups with column names}
}
\description{
Collapses columns to specified groups
}
\keyword{internal}

