% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_ascendant}
\alias{get_ascendant}
\title{Get ascendant}
\usage{
get_ascendant(x, element_id)
}
\arguments{
\item{x}{a statement object}

\item{element_id}{elements}
}
\description{
Gets common ascendant from elements
}
\keyword{internal}

