% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_descendants}
\alias{get_descendants}
\title{Get descendands}
\usage{
get_descendants(x, element_id, all = FALSE)
}
\arguments{
\item{x}{a statement object}

\item{element_id}{element (or vector of elements) from statement hierarchy}
}
\description{
Gets all descendand terminal elements from calculation tree
}
\keyword{internal}

