% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_elements}
\alias{get_elements}
\title{Get the terminating nodes of the calculation hierarchy}
\usage{
get_elements(x, parent_id = NULL, all = TRUE)
}
\arguments{
\item{x}{A statement object}

\item{parent_id}{used as filter if defined}

\item{all}{if FALSE only terminal elements from the hierarchy will be returned}
}
\description{
Get the terminating nodes of the calculation hierarchy
}
\seealso{
\code{\link{get_elements}} and \code{\link{expose}}
}

