% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_elements_h}
\alias{get_elements_h}
\title{Get elements hierarchy}
\usage{
get_elements_h(elements)
}
\arguments{
\item{elements}{data.frame with elementId, parentId and order columns}
}
\description{
Get elements hierarchy
}
\details{
Add level and hierarchycal id columns
}
\keyword{internal}

