% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_parent}
\alias{get_parent}
\title{Get parent}
\usage{
get_parent(x, element_id)
}
\arguments{
\item{x}{a statement object}

\item{element_id}{element id}
}
\description{
Gets a parent element id
}
\keyword{internal}

