% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_relations}
\alias{get_relations}
\title{Get calculation hierarchy of the statement elements}
\usage{
get_relations(x)
}
\arguments{
\item{x}{a statement object}
}
\description{
Get calculation hierarchy of the statement elements
}
\seealso{
\code{\link{get_elements}}, \code{\link{xbrl_get_statements}}
}

