% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{\%without\%}
\alias{\%without\%}
\title{Used in expose to sum the concepts under e1 without e2}
\usage{
e1 \%without\% e2
}
\arguments{
\item{e1}{element group}

\item{e2}{element group not to be included}
}
\description{
Used in expose to sum the concepts under e1 without e2
}
\keyword{internal}

