% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{join}
\alias{join}
\title{Join two statements}
\usage{
join(x, y)
}
\arguments{
\item{x}{a statement object}

\item{y}{a statement object}
}
\value{
a statement object
}
\description{
Use join for different statements in the same period.
Use merge for merging two periods of the same type of statement
}
\seealso{
\link{merge.statement} for merging two statements
}

