% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{merge.elements}
\alias{merge.elements}
\title{Merge statement elements object}
\usage{
\method{merge}{elements}(x, y, ...)
}
\arguments{
\item{x}{elements object}

\item{y}{elements object}

\item{...}{ignored}
}
\description{
Merge statement elements object
}
\keyword{internal}

