% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{merge.statement}
\alias{merge.statement}
\title{Merge two financial statements}
\usage{
\method{merge}{statement}(x, y, replace_na = TRUE, ...)
}
\arguments{
\item{x}{statement object}

\item{y}{statement object}

\item{replace_na}{(boolean) replace NAs with zeros}

\item{...}{further arguments passed to or from other methods}
}
\value{
statement object
}
\description{
Merge two statements from different time periods.
}
\details{
Since statements are basically data.frames the functions are similar. Except:
\itemize{
\item new statement elements are allways union of input elements,
  if taxonomy changes new and old elements are visible in all periods 
\item missing values are set to 0 not to NA as merge would normally do 
\item hierarchy is merged from both statements hierarchies
\item rows are treated as duplicated based on endDate and the row of x is 
allways considered as duplicate (so y should allways be the later statement)
}
}

