% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{merge.statements}
\alias{merge.statements}
\title{Merge two lists of statements}
\usage{
\method{merge}{statements}(x, y, replace_na = TRUE, ...)
}
\arguments{
\item{x}{statements object}

\item{y}{statements object}

\item{replace_na}{(boolean) replace NAs with zeros}

\item{...}{further arguments passed to or from other methods}
}
\value{
statements object
}
\description{
Merge two lists of statements
}
\details{
Merges all statements in x with all statements in y
}
\seealso{
\link{merge.statement} for merging two statements
}

