% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{other}
\alias{other}
\title{Used in expose to sum the concepts not already used}
\usage{
other(element = NULL)
}
\arguments{
\item{element}{element from concept hierarchy}
}
\description{
Used in expose to sum the concepts not already used
}
\keyword{internal}

