% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr_plot.R
\name{plot_double_stacked_bar}
\alias{plot_double_stacked_bar}
\title{Plot double stacked bar}
\usage{
plot_double_stacked_bar(x, by_date = TRUE, is_diff = FALSE,
  dif_labels = NULL, ...)
}
\arguments{
\item{x}{a statement object with}

\item{by_date}{if FALSE facets will be grouped by first element level}

\item{is_diff}{if TRUE data is rearanged to support negative values}

\item{dif_labels}{vector with two new labels in case of diff mode}

\item{...}{other attributes}
}
\description{
Plot double stacked bar
}

