% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_plot.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall}
\usage{
plot_waterfall(x, date = NULL, ...)
}
\arguments{
\item{x}{a statement object (usually income statement)}

\item{date}{end date of selected period}

\item{...}{other attributes}
}
\description{
Plots income statement as waterfall bar chart
}
\details{
Requires ggplot2.
}

