% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr_print.R
\name{print.check}
\alias{print.check}
\title{Print check object}
\usage{
\method{print}{check}(x, ...)
}
\arguments{
\item{x}{check object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print check object
}
\keyword{internal}

