% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_print.R
\name{print.elements}
\alias{print.elements}
\title{Print elements object}
\usage{
\method{print}{elements}(x, descriptions = FALSE, ...)
}
\arguments{
\item{x}{elements object}

\item{descriptions}{prints labels or element IDs}
}
\description{
Print elements object
}
\keyword{internal}

