% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_print.R
\name{print.statements}
\alias{print.statements}
\title{Print a statements object}
\usage{
\method{print}{statements}(x, ...)
}
\arguments{
\item{x}{a statements object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a statements object
}
\seealso{
xbrl_get_statements
}
\keyword{internal}

