% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_print.R
\name{print.xbrl_relations}
\alias{print.xbrl_relations}
\title{Print relations object}
\usage{
\method{print}{xbrl_relations}(x, ...)
}
\arguments{
\item{x}{relations object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print relations object
}
\seealso{
relations
}
\keyword{internal}

