% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_print.R
\name{print.xbrl_vars}
\alias{print.xbrl_vars}
\title{Print a parsed XBRL document}
\usage{
\method{print}{xbrl_vars}(x, ...)
}
\arguments{
\item{x}{XBRL data: a list of data frames as returned from XBRL::xbrlDoAll or xbrl_parse_min}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a parsed XBRL document
}
\seealso{
xbrl_parse_min, xbrl_get_statements
}
\keyword{internal}

