% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{proportional}
\alias{proportional}
\title{Proportional values}
\usage{
proportional(x, digits = NULL)
}
\arguments{
\item{x}{a statement object}

\item{digits}{if specified number of digits to round the result}
}
\description{
Every value in the financial statement is divided by topmost parent's value
}

