% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{reshape_long}
\alias{reshape_long}
\title{Reshape to "long" format}
\usage{
reshape_long(x, levels = NULL)
}
\description{
Reshapes statement object to a data frame with one value column
 and dimension columns (endDate, elementId and parentId). 
 @param x a statement object
 @param levels if defined only elements from specified levels will be included
 @export
}

