% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{reshape_table}
\alias{reshape_table}
\title{Reshape statement data to table format}
\usage{
reshape_table(x, decimals = TRUE, simple = FALSE)
}
\arguments{
\item{x}{a statement object}

\item{decimals}{return values with decimals}
}
\description{
Transposes data to "print-out" format
}
\keyword{internal}

