% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{xbrl_create_data}
\alias{xbrl_create_data}
\title{Function to create package data included in the package}
\usage{
xbrl_create_data()
}
\description{
Function to create package data included in the package
}
\details{
xbrlDoAll creates > 5Mb list - we need only 2Mb
}
\keyword{internal}

