% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{xbrl_get_data}
\alias{xbrl_get_data}
\title{Get a statement from data (data for specified elements)}
\usage{
xbrl_get_data(elements, xbrl_vars, complete_only = FALSE,
  complete_first = TRUE, basic_contexts = TRUE)
}
\arguments{
\item{elements}{elements object}

\item{xbrl_vars}{XBRL data}

\item{complete_only}{just the rows without NA}

\item{complete_first}{just the rows without NA in first column}

\item{basic_contexts}{in case of duplicated periods, get only basic contexts}
}
\description{
Get a statement from data (data for specified elements)
}
\keyword{internal}

