% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{xbrl_get_relations}
\alias{xbrl_get_relations}
\title{Get relations from XBRL calculation link base}
\usage{
xbrl_get_relations(xbrl_vars, role_id, lbase = "calculation")
}
\arguments{
\item{xbrl_vars}{XBRL data (list of dataframes)}

\item{role_id}{id of role (usually of type statement)}

\item{lbase}{link base (default is calculation)}
}
\description{
Get relations from XBRL calculation link base
}
\keyword{internal}

