% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{xbrl_get_statement_ids}
\alias{xbrl_get_statement_ids}
\title{Get a vector of statement IDs}
\usage{
xbrl_get_statement_ids(xbrl_vars)
}
\arguments{
\item{xbrl_vars}{XBRL data}
}
\description{
Get a vector of statement IDs
}
\keyword{internal}

