% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{calculate}
\alias{calculate}
\title{Calculate formulas}
\usage{
calculate(x, ..., digits = NULL, decimals = NULL)
}
\arguments{
\item{x}{a statement object}

\item{...}{list of formulas}

\item{digits}{if specified the result will be rounded according to number of digits}

\item{decimals}{if specified the result will be multiplied by 10^decimals}
}
\value{
data frame with date and specified columns
}
\description{
Calculate formulas
}
\examples{
\dontrun{

balance_sheet \%>\% calculate(

  current_ratio = AssetsCurrent / LiabilitiesCurrent,

  quick_ratio =
    ( CashAndCashEquivalentsAtCarryingValue +
        AvailableForSaleSecuritiesCurrent +
        AccountsReceivableNetCurrent
      ) / LiabilitiesCurrent
)
}
}
\seealso{
\code{\link{calculation}}
}

