% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{calculation}
\alias{calculation}
\title{Define calculation}
\usage{
calculation(..., x = NULL)
}
\arguments{
\item{...}{formulas}

\item{x}{not used}
}
\description{
Define calculation
}
\examples{
\dontrun{

profit_margins <- calculation(

 Gross_Margin =
   (SalesRevenueNet -  CostOfGoodsAndServicesSold) / SalesRevenueNet,
 Operating_Margin =
   OperatingIncomeLoss / SalesRevenueNet,
 Net_Margin =
   NetIncomeLoss / SalesRevenueNet,

 digits = 2
)

income_statement \%>\% do_calculation(profit_margins)
}
}
\seealso{
\code{\link{do_calculation}}
}

