% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{calculation}
\alias{calculation}
\title{Define calculation}
\usage{
calculation(...)
}
\arguments{
\item{...}{formulas}
}
\description{
Define calculation
}
\examples{
\dontrun{

profit_margins <- calculation(
 Gross_Margin = (SalesRevenueNet - CostOfGoodsAndServicesSold) / SalesRevenueNet,
 Operating_Margin = OperatingIncomeLoss / SalesRevenueNet,
 Net_Margin = NetIncomeLoss / SalesRevenueNet
)

income_statement \%>\% calculate(calculations = profit_margins)
}
}
\seealso{
\code{\link{calculate}}
}

