% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{check_statement}
\alias{check_statement}
\title{Check statement}
\usage{
check_statement(statement, element_id = NULL)
}
\arguments{
\item{statement}{statement object}

\item{element_id}{element from hierarchy where to perform calculation (if specified)}
}
\value{
check object with calculated and original values
}
\description{
Checks statement calculation consistency
}

