% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{do_calculation}
\alias{do_calculation}
\title{Run a calculation}
\usage{
do_calculation(x, ...)
}
\arguments{
\item{x}{statement object}

\item{...}{one or more calculations}
}
\value{
a data frame or a list of data frames
}
\description{
Calculations can be defined with \code{\link{calculation}}
 function.
 With \code{\link{do_calculation}} the calculations get calculated on a
 specified statement.
 Resulting data frame (or data frames if there are several calculations) has
 a date column and the columns defined with calculations.
}
\seealso{
\code{\link{calculation}}
}

