% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{expose}
\alias{expose}
\title{Expose financial sheet values}
\usage{
expose(x, ..., e_list = NULL)
}
\arguments{
\item{x}{a statement object}

\item{...}{expressions to expose values}

\item{e_list}{expressions to expose values}
}
\description{
Simplifies statement to 2-level hierarchy.
Elements are defined by list of element vectors.
}
\examples{
\dontrun{
expose(balance_sheet,

  # Assets
  `Current Assets` = "AssetsCurrent",
  `Noncurrent Assets` = other("Assets"),
  # Liabilites and equity
  `Current Liabilities` = "LiabilitiesCurrent",
  `Noncurrent Liabilities` = other(c("Liabilities", "CommitmentsAndContingencies")),
  `Stockholders Equity` = "StockholdersEquity"
)
}
}

