% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_elements}
\alias{get_elements}
\title{Get statement elements and the calculation hierarchy}
\usage{
get_elements(x, parent_id = NULL, all = TRUE)
}
\arguments{
\item{x}{A statement object}

\item{parent_id}{used as filter if defined}

\item{all}{if FALSE only terminal elements from the hierarchy will be returned}
}
\description{
Get statement elements and the calculation hierarchy
}
\seealso{
\code{\link{calculate}}
}

