% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{get_elements}
\alias{get_elements}
\title{Get the terminating nodes of the calculation hierarchy}
\usage{
get_elements(x, parent_id = NULL, as_data_frame = FALSE)
}
\arguments{
\item{x}{A statement object}

\item{parent_id}{Ascendant element - used as filter if defined}

\item{as_data_frame}{If TRUE, the return value is in a data frame format (defaults to FALSE)}
}
\description{
Get the terminating nodes of the calculation hierarchy
}
\seealso{
\code{\link{get_relations}} and \code{\link{expose}}
}

