% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{\%without\%}
\alias{\%without\%}
\title{Without operator}
\usage{
a \%without\% b
}
\arguments{
\item{a}{element id from element hierarchy}

\item{b}{element id from element hierarchy}
}
\value{
Lazy object with function to select statement elements in a witout elements in b
}
\description{
Used inside fold function to select elements in a without elements in b
}
\keyword{internal}

