% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{other}
\alias{other}
\title{Other elements}
\usage{
other(...)
}
\arguments{
\item{...}{element IDs from element hierarchy}
}
\value{
Lazy object with function to select elements from statement x
}
\description{
Used inside expose function to select all other elements
}
\keyword{internal}

