% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_plot.R
\name{plot_double_stacked_bar}
\alias{plot_double_stacked_bar}
\title{Plot double stacked bar}
\usage{
plot_double_stacked_bar(x, by_date = FALSE, is_diff = FALSE,
  dif_labels = NULL)
}
\arguments{
\item{x}{a list with two data frames (sides)}

\item{by_date}{if TRUE, facets will be grouped by date}

\item{is_diff}{if TRUE data is rearanged to support negative values}

\item{dif_labels}{vector with two new labels in case of diff mode}
}
\description{
Plot double stacked bar
}

