% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_print.R
\name{print.statement}
\alias{print.statement}
\title{Print a statement object}
\usage{
\method{print}{statement}(x, descriptions = FALSE, html = FALSE,
  big.mark = "", dateFormat = NULL, ...)
}
\arguments{
\item{x}{a statement object}

\item{descriptions}{if TRUE labels are printed instead of element IDs}

\item{html}{print in html format (requires htmlTable)}

\item{big.mark}{big mark to format numbers}

\item{dateFormat}{format string for format dates}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a statement object
}

