% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr_print.R
\name{print.statement}
\alias{print.statement}
\title{Print a statement object}
\usage{
\method{print}{statement}(x, ...)
}
\arguments{
\item{x}{a statement object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a statement object
}

