% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\docType{data}
\name{xbrl_data_aapl2013}
\alias{xbrl_data_aapl2013}
\title{XBRL data from SEC archive}
\description{
parsed XBRL files from
   http://edgar.sec.gov/Archives/edgar/data/320193/000119312514383437/aapl-20140927.xml
}

