% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finstr.R
\name{xbrl_get_data}
\alias{xbrl_get_data}
\title{Get a statement from data (data for specified elements)}
\usage{
xbrl_get_data(elements, xbrl_vars, complete_only = TRUE)
}
\arguments{
\item{elements}{elements object}

\item{xbrl_vars}{XBRL data}

\item{complete_only}{just the rows without NA}
}
\description{
Get a statement from data (data for specified elements)
}
\keyword{internal}

