% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finstr.R
\name{xbrl_get_statements}
\alias{xbrl_get_statements}
\title{Get a financial statements object from XBRL}
\usage{
xbrl_get_statements(xbrl_vars, rm_prefix = "us-gaap_")
}
\arguments{
\item{xbrl_vars}{a XBRL parsed object}

\item{rm_prefix}{a prefix to remove from element names}
}
\value{
A statements object
}
\description{
Get a financial statements object from XBRL
}
\examples{
\dontrun{

# parse XBRL to XBRL data and get statements:
xbrl_url <- "http://edgar.sec.gov/Archives/edgar/data/320193/000119312514383437/aapl-20140927.xml"
xbrl_data <- xbrl_parse_min(xbrl_url)
st1 <- xbrl_get_statements(xbrl_data)

# get statements directly from a url:
xbrl_url <- "http://edgar.sec.gov/Archives/edgar/data/320193/000119312514383437/aapl-20140927.xml"
st1 <- xbrl_get_statements(xbrl_url)
}

}
\seealso{
\link{finstr}
}

