% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finstr.R
\name{xbrl_parse_min}
\alias{xbrl_parse_min}
\title{Parse XBRL}
\usage{
xbrl_parse_min(xbrl_file, parse_labels = FALSE)
}
\arguments{
\item{xbrl_file}{file path or url to XBRL file}

\item{parse_labels}{if element labels should be parsed (default to FALSE)}
}
\value{
xbrl_data object
}
\description{
Parse XBRL files
}
\details{
xbrl_parse_min uses primitives from
 \link[XBRL]{XBRL-package} to parse only necessary information needed to
 convert XBRL to statement object with \code{\link{xbrl_get_statements}}
 function.
 It is slightly faster than \code{\link{xbrlDoAll}} but it does not parse
 all the data.
 It parses only roles, elements, contexts, facts and calculation base link.
}
\seealso{
\code{\link{xbrl_get_statements}}
}

